+++ src/Helpers.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use AmdadulHaq\Setting\Models\Setting;
+use AmdadulHaq\Setting\Setting;
 
 if (! function_exists('setting')) {
     function setting(): Setting
     {
-        return new Setting();
+        return resolve(Setting::class);
     }
 }